+++ test/e2e/onFID-test.js
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-const assert = require('assert');
-const {beaconCountIs, clearBeacons, getBeacons} = require('../utils/beacons.js');
-const {browserSupportsEntry} = require('../utils/browserSupportsEntry.js');
-const {stubForwardBack} = require('../utils/stubForwardBack.js');
-const {stubVisibilityChange} = require('../utils/stubVisibilityChange.js');
+import assert from 'assert';
+import {beaconCountIs, clearBeacons, getBeacons} from '../utils/beacons.js';
+import {browserSupportsEntry} from '../utils/browserSupportsEntry.js';
+import {domReadyState} from '../utils/domReadyState.js';
+import {stubForwardBack} from '../utils/stubForwardBack.js';
+import {stubVisibilityChange} from '../utils/stubVisibilityChange.js';
 
 
-describe('getFID()', async function() {
+describe('onFID()', async function() {
   // Retry all tests in this suite up to 2 times.
   this.retries(2);
 
@@ -47,10 +48,12 @@ describe('getFID()', async function() {
 
     const [fid] = await getBeacons();
     assert(fid.value >= 0);
-    assert(fid.id.match(/^v2-\d+-\d+$/));
+    assert(fid.id.match(/^v3-\d+-\d+$/));
     assert.strictEqual(fid.name, 'FID');
     assert.strictEqual(fid.value, fid.delta);
-    assert.strictEqual(fid.entries[0].name, 'mousedown');
+    assert.strictEqual(fid.rating, 'good');
+    assert.match(fid.navigationType, /navigate|reload/);
+    assert.match(fid.entries[0].name, /(mouse|pointer)down/);
   });
 
   it('does not report if the browser does not support FID and the polyfill is not used', async function() {
@@ -97,10 +100,12 @@ describe('getFID()', async function() {
     const [fid] = await getBeacons();
 
     assert(fid.value >= 0);
-    assert(fid.id.match(/^v2-\d+-\d+$/));
+    assert(fid.id.match(/^v3-\d+-\d+$/));
     assert.strictEqual(fid.name, 'FID');
     assert.strictEqual(fid.value, fid.delta);
-    assert.strictEqual(fid.entries[0].name, 'mousedown');
+    assert.strictEqual(fid.rating, 'good');
+    assert.match(fid.navigationType, /navigate|reload/);
+    assert.match(fid.entries[0].name, /(mouse|pointer)down/);
     if (browserSupportsFID) {
       assert('duration' in fid.entries[0]);
     } else {
@@ -114,6 +119,7 @@ describe('getFID()', async function() {
     if (browser.capabilities.browserName === 'Safari') this.skip();
 
     await browser.url('/test/fid?hidden=1');
+    await domReadyState('interactive');
 
     await stubVisibilityChange('visible');
 
@@ -162,10 +168,12 @@ describe('getFID()', async function() {
 
     const [fid1] = await getBeacons();
     assert(fid1.value >= 0);
-    assert(fid1.id.match(/^v2-\d+-\d+$/));
+    assert(fid1.id.match(/^v3-\d+-\d+$/));
     assert.strictEqual(fid1.name, 'FID');
     assert.strictEqual(fid1.value, fid1.delta);
-    assert.strictEqual(fid1.entries[0].name, 'mousedown');
+    assert.strictEqual(fid1.rating, 'good');
+    assert.match(fid1.navigationType, /navigate|reload/);
+    assert.match(fid1.entries[0].name, /(mouse|pointer)down/);
 
     await clearBeacons();
     await stubForwardBack();
@@ -177,11 +185,72 @@ describe('getFID()', async function() {
 
     const [fid2] = await getBeacons();
     assert(fid2.value >= 0);
-    assert(fid2.id.match(/^v2-\d+-\d+$/));
+    assert(fid2.id.match(/^v3-\d+-\d+$/));
     assert(fid1.id !== fid2.id);
     assert.strictEqual(fid2.name, 'FID');
+    assert.strictEqual(fid2.rating, 'good');
     assert.strictEqual(fid2.value, fid2.delta);
-    assert.strictEqual(fid2.entries[0].name, 'mousedown');
+    assert.strictEqual(fid2.navigationType, 'back-forward-cache');
+    assert.match(fid2.entries[0].name, /(mouse|pointer)down/);
+  });
+
+  describe('attribution', function() {
+    it('includes attribution data on the metric object', async function() {
+      if (!browserSupportsFID) this.skip();
+
+      await browser.url('/test/fid?attribution=1');
+
+      // Click on the <h1>.
+      const h1 = await $('h1');
+      await h1.click();
+
+      await beaconCountIs(1);
+
+      const [fid] = await getBeacons();
+      assert(fid.value >= 0);
+      assert(fid.id.match(/^v3-\d+-\d+$/));
+      assert.strictEqual(fid.name, 'FID');
+      assert.strictEqual(fid.value, fid.delta);
+      assert.strictEqual(fid.rating, 'good');
+      assert.match(fid.navigationType, /navigate|reload/);
+      assert.match(fid.entries[0].name, /(mouse|pointer)down/);
+
+      // This value is frequently not set in Chrome for some reason,
+      // so just check that it's a string.
+      assert(typeof fid.attribution.eventTarget === 'string');
+      assert.equal(fid.attribution.eventTime, fid.entries[0].startTime);
+      assert.equal(fid.attribution.eventType, fid.entries[0].name);
+      assert.deepEqual(fid.attribution.eventEntry, fid.entries[0]);
+      assert.equal(fid.attribution.loadState, 'complete');
+    });
+
+    it('reports the domReadyState when input occurred', async function() {
+      if (!browserSupportsFID) this.skip();
+
+      await browser.url('/test/fid?attribution=1&delayDCL=1000');
+
+      // Click on the <h1>.
+      const h1 = await $('h1');
+      await h1.click();
+
+      await beaconCountIs(1);
+
+      const [fid1] = await getBeacons();
+      assert.equal(fid1.attribution.loadState, 'dom-interactive');
+
+      await clearBeacons();
+
+      await browser.url('/test/fid?attribution=1&delayResponse=1000');
+
+      // Click on the <h1>.
+      const p = await $('p');
+      await p.click();
+
+      await beaconCountIs(1);
+
+      const [fid2] = await getBeacons();
+      assert.equal(fid2.attribution.loadState, 'loading');
+    });
   });
 });