+++ test/utils/stubForwardBack.js
@@ -20,34 +20,8 @@
  * attribute (to prevent painting) and dispatches a `visibilitychange` event.
  * @return {Promise<void>}
  */
-function stubForwardBack(visibilityStateAfterRestore) {
+export function stubForwardBack(visibilityStateAfterRestore) {
   return browser.executeAsync((visibilityStateAfterRestore, done) => {
-    window.dispatchEvent(new PageTransitionEvent('pagehide', {
-      persisted: true,
-    }));
-    Object.defineProperty(document, 'visibilityState', {
-      value: 'hidden',
-      configurable: true,
-    });
-    document.body.hidden = true;
-    document.dispatchEvent(new Event('visibilitychange'));
-
-    requestAnimationFrame(() => {
-      requestAnimationFrame(() => {
-        if (visibilityStateAfterRestore !== 'hidden') {
-          delete document.visibilityState;
-          document.body.hidden = false;
-        }
-        document.dispatchEvent(new Event('visibilitychange'));
-        window.dispatchEvent(new PageTransitionEvent('pageshow', {
-          persisted: true,
-        }));
-        done();
-      });
-    });
+    self.__stubForwardBack(visibilityStateAfterRestore).then(done);
   }, visibilityStateAfterRestore);
 }
-
-module.exports = {
-  stubForwardBack,
-};