+++ test/utils/stubVisibilityChange.js
@@ -20,22 +20,8 @@
  * attribute (to prevent painting) and dispatches a `visibilitychange` event.
  * @return {Promise<void>}
  */
-function stubVisibilityChange(visibilityState) {
+export function stubVisibilityChange(visibilityState) {
   return browser.execute((visibilityState) => {
-    if (visibilityState === 'hidden') {
-      Object.defineProperty(document, 'visibilityState', {
-        value: visibilityState,
-        configurable: true,
-      });
-      document.body.hidden = true;
-    } else {
-      delete document.visibilityState;
-      document.body.hidden = false;
-    }
-    document.dispatchEvent(new Event('visibilitychange'));
+    self.__stubVisibilityChange(visibilityState);
   }, visibilityState);
 }
-
-module.exports = {
-  stubVisibilityChange,
-};